HW 12.

+++ src/main/java/TrianglePage/TriangleArea.java
@@ -2,8 +2,8 @@ package TrianglePage;
 
 public class TriangleArea {
 
-    public static double calculate(double a, double b, double c) {
-        double pp = (a + b + c) / 2D;  // полупериметр
+    public static double calculate(int a, int b, int c) {
+        double pp = (a + b + c) / 2.0;  // полупериметр
         return Math.sqrt(pp * (pp - a) * (pp - b) * (pp - c)); // формулa Герона
 
     }
